+++ bindings/nodejs/types/client/bridge/index.ts
@@ -1,0 +1,132 @@
+import type {
+    __GetInfoMethod__,
+    __GetBasicOutputIdsMethod__,
+    __GetOutputMethod__,
+    __GetOutputsMethod__,
+    __GenerateAddressesMethod__,
+    __PostBlockMethod__,
+    __BuildAndPostBlockMethod__,
+    __GetTipsMethod__,
+    __GetNetworkInfoMethod__,
+    __GetBlockMethod__,
+    __GetBlockMetadataMethod__,
+    __FindInputsMethod__,
+    __FindOutputsMethod__,
+    __PrepareTransactionMethod__,
+    __SignTransactionMethod__,
+    __PostBlockPayloadMethod__,
+    __GetNodeMethod__,
+    __GetNetworkIdMethod__,
+    __GetBech32HrpMethod__,
+    __GetMinPowScoreMethod__,
+    __GetTipsIntervalMethod__,
+    __GetProtocolParametersMethod__,
+    __GetLocalPowMethod__,
+    __GetFallbackToLocalPowMethod__,
+    __GetHealthMethod__,
+    __GetNodeInfoMethod__,
+    __GetPeersMethod__,
+    __PostBlockRawMethod__,
+    __GetBlockRawMethod__,
+    __GetReceiptsMethod__,
+    __GetReceiptsMigratedAtMethod__,
+    __GetTreasuryMethod__,
+    __GetIncludedBlockMethod__,
+    __GetIncludedBlockMetadataMethod__,
+    __HexToBech32Method__,
+    __AliasIdToBech32Method__,
+    __NftIdToBech32Method__,
+    __HexPublicKeyToBech32AddressMethod__,
+    __AliasOutputIdsMethod__,
+    __AliasOutputIdMethod__,
+    __NftOutputIdsMethod__,
+    __NftOutputIdMethod__,
+    __FoundryOutputIdsMethod__,
+    __FoundryOutputIdMethod__,
+    __TryGetOutputsMethod__,
+    __FindBlocksMethod__,
+    __RetryMethod__,
+    __RetryUntilIncludedMethod__,
+    __ConsolidateFundsMethod__,
+    __ReattachMethod__,
+    __ReattachUncheckedMethod__,
+    __PromoteMethod__,
+    __PromoteUncheckedMethod__,
+    __UnhealthyNodesMethod__,
+    __GetMilestoneByIdMethod__,
+    __GetUtxoChangesByIdMethod__,
+    __GetMilestoneByIndexMethod__,
+    __GetUtxoChangesByIndexMethod__,
+    __BuildBasicOutputMethod__,
+    __BuildAliasOutputMethod__,
+    __BuildFoundryOutputMethod__,
+    __BuildNftOutputMethod__,
+    __ClearListenersMethod__,
+    __SignatureUnlockMethod__,
+} from './client';
+
+export type __ClientMethods__ =
+    | __GetInfoMethod__
+    | __GetOutputMethod__
+    | __GetBasicOutputIdsMethod__
+    | __GetOutputsMethod__
+    | __GenerateAddressesMethod__
+    | __PostBlockMethod__
+    | __BuildAndPostBlockMethod__
+    | __GetTipsMethod__
+    | __GetNetworkInfoMethod__
+    | __GetBlockMethod__
+    | __GetBlockMetadataMethod__
+    | __FindInputsMethod__
+    | __FindOutputsMethod__
+    | __PrepareTransactionMethod__
+    | __SignTransactionMethod__
+    | __SignatureUnlockMethod__
+    | __PostBlockPayloadMethod__
+    | __GetNodeMethod__
+    | __GetNetworkIdMethod__
+    | __GetBech32HrpMethod__
+    | __GetMinPowScoreMethod__
+    | __GetTipsIntervalMethod__
+    | __GetProtocolParametersMethod__
+    | __GetLocalPowMethod__
+    | __GetFallbackToLocalPowMethod__
+    | __GetHealthMethod__
+    | __GetNodeInfoMethod__
+    | __GetPeersMethod__
+    | __PostBlockRawMethod__
+    | __GetBlockRawMethod__
+    | __GetMilestoneByIdMethod__
+    | __GetUtxoChangesByIdMethod__
+    | __GetMilestoneByIndexMethod__
+    | __GetUtxoChangesByIndexMethod__
+    | __GetReceiptsMethod__
+    | __GetReceiptsMigratedAtMethod__
+    | __GetTreasuryMethod__
+    | __GetIncludedBlockMethod__
+    | __GetIncludedBlockMetadataMethod__
+    | __HexToBech32Method__
+    | __AliasIdToBech32Method__
+    | __NftIdToBech32Method__
+    | __HexPublicKeyToBech32AddressMethod__
+    | __AliasOutputIdsMethod__
+    | __AliasOutputIdMethod__
+    | __NftOutputIdsMethod__
+    | __NftOutputIdMethod__
+    | __FoundryOutputIdsMethod__
+    | __FoundryOutputIdMethod__
+    | __TryGetOutputsMethod__
+    | __FindBlocksMethod__
+    | __RetryMethod__
+    | __RetryUntilIncludedMethod__
+    | __ConsolidateFundsMethod__
+    | __ReattachMethod__
+    | __ReattachUncheckedMethod__
+    | __PromoteMethod__
+    | __PromoteUncheckedMethod__
+    | __UnhealthyNodesMethod__
+    | __BuildBasicOutputMethod__
+    | __BuildAliasOutputMethod__
+    | __BuildFoundryOutputMethod__
+    | __BuildNftOutputMethod__
+    | __ClearListenersMethod__;