+++ sdk/src/client/bindings/nodejs/jest.config.js
@@ -1,9 +1,0 @@
-/** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
-module.exports = {
-    preset: 'ts-jest',
-    testEnvironment: 'node',
-    testMatch: ['<rootDir>/test/**/*.(test|spec).ts'],
-    moduleNameMapper: {
-        'index.node': '<rootDir>/build/Release/index.node',
-    },
-};