+++ sdk/tests/wallet/claim_outputs.rs
@@ -16,8 +16,8 @@ use crate::wallet::common::{create_accounts_with_funds, make_wallet, setup, tear
 
 #[ignore]
 #[tokio::test]
-async fn claim_2_basic_outputs() -> Result<()> {
-    let storage_path = "test-storage/claim_2_basic_outputs";
+async fn claim_2_basic_micro_outputs() -> Result<()> {
+    let storage_path = "test-storage/claim_2_basic_micro_outputs";
     setup(storage_path)?;
 
     let wallet = make_wallet(storage_path, None, None).await?;
@@ -63,6 +63,60 @@ async fn claim_2_basic_outputs() -> Result<()> {
     assert_eq!(
         balance.base_coin().available(),
         base_coin_amount_before_claiming + 2 * micro_amount
+    );
+
+    tear_down(storage_path)
+}
+
+#[ignore]
+#[tokio::test]
+async fn claim_1_of_2_basic_outputs() -> Result<()> {
+    let storage_path = "test-storage/claim_1_of_2_basic_outputs";
+    setup(storage_path)?;
+
+    let wallet = make_wallet(storage_path, None, None).await?;
+
+    let accounts = create_accounts_with_funds(&wallet, 2).await?;
+
+    let amount = 10;
+    let tx = accounts[1]
+        .send_amount(
+            vec![
+                AddressWithAmount::new(accounts[0].addresses().await?[0].address().to_string(), amount),
+                AddressWithAmount::new(accounts[0].addresses().await?[0].address().to_string(), 0),
+            ],
+            TransactionOptions {
+                allow_micro_amount: true,
+                ..Default::default()
+            },
+        )
+        .await?;
+
+    accounts[1]
+        .retry_transaction_until_included(&tx.transaction_id, None, None)
+        .await?;
+
+    // Claim with account 0
+    let balance = accounts[0].sync(None).await.unwrap();
+    assert_eq!(balance.potentially_locked_outputs().len(), 2);
+    let base_coin_amount_before_claiming = balance.base_coin().available();
+
+    let tx = accounts[0]
+        .claim_outputs(
+            accounts[0]
+                .get_unlockable_outputs_with_additional_unlock_conditions(OutputsToClaim::Amount)
+                .await?,
+        )
+        .await?;
+    accounts[0]
+        .retry_transaction_until_included(&tx.transaction_id, None, None)
+        .await?;
+
+    let balance = accounts[0].sync(None).await.unwrap();
+    assert_eq!(balance.potentially_locked_outputs().len(), 1);
+    assert_eq!(
+        balance.base_coin().available(),
+        base_coin_amount_before_claiming + amount
     );
 
     tear_down(storage_path)
@@ -292,7 +346,7 @@ async fn claim_2_native_tokens_no_outputs_in_claim_account() -> Result<()> {
     let tx = account_0
         .send(
             vec![
-                BasicOutputBuilder::new_with_minimum_storage_deposit(rent_structure.clone())
+                BasicOutputBuilder::new_with_minimum_storage_deposit(rent_structure)
                     .add_unlock_condition(AddressUnlockCondition::new(
                         *account_1.addresses().await?[0].address().as_ref(),
                     ))
@@ -302,7 +356,7 @@ async fn claim_2_native_tokens_no_outputs_in_claim_account() -> Result<()> {
                     )?)
                     .add_native_token(NativeToken::new(mint_tx_0.token_id, native_token_amount)?)
                     .finish_output(token_supply)?,
-                BasicOutputBuilder::new_with_minimum_storage_deposit(rent_structure.clone())
+                BasicOutputBuilder::new_with_minimum_storage_deposit(rent_structure)
                     .add_unlock_condition(AddressUnlockCondition::new(
                         *account_1.addresses().await?[0].address().as_ref(),
                     ))